windows/requirements.txt
Pillow
pynput==1.7.6
pywin32